fix: fixed issues thrown by sonar
