+++ Seeder/Seeder_Application/src/theme/index.tsx
@@ -1,5 +1,4 @@
 import { createTheme } from '@mui/material/styles';
-import '@mui/material/styles';
 
 declare module '@mui/material/styles' {
